Move loader and events to plugin_system
